+++ packages/sui-bundler/webpack.config.prod.js
@@ -7,7 +7,6 @@ const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
 const path = require('path')
 const ScriptExtHtmlWebpackPlugin = require('script-ext-html-webpack-plugin')
 
-const {GenerateSW} = require('workbox-webpack-plugin')
 const minifyJs = require('./shared/minify-js')
 const webpack = require('webpack')
 const definePlugin = require('./shared/define')
@@ -18,11 +17,6 @@ const crypto = require('crypto')
 const zlib = require('zlib')
 const hasBrotliSupport = Boolean(zlib.brotliCompress)
 
-const {
-  navigateFallbackDenylist,
-  navigateFallback,
-  runtimeCaching
-} = require('./shared/precache')
 const {when, cleanList, envVars, MAIN_ENTRY_POINT, config} = require('./shared')
 const {sourceMap} = require('./shared/config')
 const parseAlias = require('./shared/parse-alias')
@@ -194,27 +188,6 @@ module.exports = {
     new ManifestPlugin({
       fileName: 'asset-manifest.json'
     }),
-    when(
-      config.offline,
-      () =>
-        new GenerateSW({
-          skipWaiting: true,
-          clientsClaim: true,
-          dontCacheBustURLsMatching: /\.\w{8}\./,
-          cleanupOutdatedCaches: true,
-          directoryIndex: config.offline.directoryIndex,
-          additionalManifestEntries: config.offline.addToCache,
-          navigateFallback: navigateFallback(
-            config.offline.fallback,
-            PUBLIC_PATH
-          ),
-          navigateFallbackDenylist: navigateFallbackDenylist(
-            config.offline.denylist
-          ),
-          runtimeCaching: runtimeCaching(config.offline.runtime),
-          exclude: [/\.map$/, /asset-manifest\.json$/, /LICENSE\.txt$/]
-        })
-    ),
     when(config.externals, () => new Externals({files: config.externals})),
     new LoaderUniversalOptionsPlugin(require('./shared/loader-options')),
     when(